+++ src/main/java/com/myprojects/demo/controllers/UserController.java
@@ -1,8 +1,6 @@
 package com.myprojects.demo.controllers;
 
-import com.myprojects.demo.dto.MovieForm;
 import com.myprojects.demo.dto.UserForm;
-import com.myprojects.demo.entities.Movie;
 import com.myprojects.demo.entities.Reaction;
 import com.myprojects.demo.entities.User;
 import com.myprojects.demo.repositories.UserRepository;
@@ -28,12 +26,6 @@ public class UserController {
         return userService.getAllUsers();
     }
 
-    @PostMapping("/{userId}/add")
-    public Movie addMovie(@PathVariable Long userId, @RequestParam MovieForm movieForm) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new EntityNotFoundException("User doesn't exist"));
-        return userService.addMovie(user, movieForm);
-    }
 
     @GetMapping("/{userId}/likes/{movieId}")
     public Reaction likeMovie(@PathVariable Long userId, @PathVariable Long movieId) {
