+++ src/main/java/com/myprojects/demo/services/UserService.java
@@ -1,6 +1,5 @@
 package com.myprojects.demo.services;
 
-import com.myprojects.demo.dto.MovieForm;
 import com.myprojects.demo.entities.Movie;
 import com.myprojects.demo.entities.Reaction;
 import com.myprojects.demo.entities.User;
@@ -32,16 +31,6 @@ public class UserService {
 
     public List<User> getAllUsers() {
         return userRepository.findAll();
-    }
-
-    public Movie addMovie(User user, MovieForm movieForm) {
-        Optional<Movie> movie = movieRepository.findByTitle(movieForm.getTitle());
-        if (movie.isPresent()) {
-            throw new InvalidInputException("Movie with title: " + movieForm.getTitle() + " already exists.");
-        }
-        Movie newMovie = new Movie(movieForm);
-        newMovie.setUploadedBy(user);
-        return movieRepository.save(newMovie);
     }
 
     @Transactional
